+++ cloud/bqfake/bqfake_test.go
@@ -14,7 +14,7 @@ import (
 	"cloud.google.com/go/bigquery"
 	"github.com/GoogleCloudPlatform/google-cloud-go-testing/bigquery/bqiface"
 	"github.com/m-lab/go/cloud/bqfake"
-	frombigquery "github.com/m-lab/go/cloud/bqfake/from-bigquery"
+	"github.com/m-lab/go/internal/frombigquery"
 	"google.golang.org/api/iterator"
 )
